backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.3",
    "asyncpg>=0.31.0",
    "authlib>=1.6.8",
    "boto3>=1.42.49",
    "celery[redis]>=5.3.1",
    "fastapi[standard]>=0.128.0",
    "greenlet>=3.3.1",
    "httpx>=0.28.1",
    "itsdangerous>=2.2.0",
    "jinja2>=3.1.6",
    "minio>=7.2.20",
    "oss2>=2.19.1",
    "passlib[bcrypt]>=1.7.4",
    "pyjwt>=2.11.0",
    "pytest-mock>=3.15.1",
    "python-json-logger>=4.0.0",
    "redis>=7.1.0",
    "sqlalchemy>=2.0.46",
]

[dependency-groups]
dev = [
    "black>=26.1.0",
    "isort>=7.0.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"
